Make modal responsive using Media queries
