--- src/lib/screenSizes.js
@@ -1,11 +1,0 @@
-// Breakpints according to material-ui library. Also in css file layout.css
-
-export const xsMax = '48em'
-
-export const smMin = '48em'
-export const smMax = '62em'
-
-export const mdMin = '62em'
-export const mdMax = '75em'
-
-export const lgMin = '75em'
