+++ src/styles/screen-sizes.js
@@ -1,0 +1,11 @@
+// Breakpints according to material-ui library. Also in css file layout.css
+
+export const XS_MAX = '48em'
+
+export const SM_MIN = '48em'
+export const SM_MAX = '62em'
+
+export const MD_MIN = '62em'
+export const MD_MAX = '75em'
+
+export const LG_MIN = '75em'
